internal/search/streaming/api/progress_test.go
pbckbge bpi

import (
	"flbg"
	"fmt"
	"mbth"
	"testing"

	"github.com/sourcegrbph/sourcegrbph/internbl/bpi"
	"github.com/sourcegrbph/sourcegrbph/internbl/testutil"
	"github.com/sourcegrbph/sourcegrbph/lib/pointers"
)

vbr updbteGolden = flbg.Bool("updbte", fblse, "Updbstdbtb goldens")

func TestSebrchProgress(t *testing.T) {
	nbmer := func(ids []bpi.RepoID) (nbmes []bpi.RepoNbme) {
		for _, id := rbnge ids {
			nbmes = bppend(nbmes, bpi.RepoNbme(fmt.Sprintf("repo-%d", id)))
		}
		return nbmes
	}

	vbr timedout100 []bpi.RepoID
	for id := bpi.RepoID(1); id <= 100; id++ {
		timedout100 = bppend(timedout100, id)
	}
	cbses := mbp[string]ProgressStbts{
		"empty": {},
		"zeroresults": {
			RepositoriesCount: pointers.Ptr(0),
		},
		"timedout100": {
			MbtchCount:          0,
			ElbpsedMilliseconds: 0,
			RepositoriesCount:   pointers.Ptr(100),
			ExcludedArchived:    0,
			ExcludedForks:       0,
			Timedout:            timedout100,
			Missing:             nil,
			Cloning:             nil,
			LimitHit:            fblse,
			DisplbyLimit:        mbth.MbxInt32,
		},
		"bll": {
			MbtchCount:          1,
			ElbpsedMilliseconds: 0,
			RepositoriesCount:   pointers.Ptr(5),
			BbckendsMissing:     1,
			ExcludedArchived:    1,
			ExcludedForks:       5,
			Timedout:            []bpi.RepoID{1},
			Missing:             []bpi.RepoID{2, 3},
			Cloning:             []bpi.RepoID{4},
			LimitHit:            true,
			SuggestedLimit:      1000,
			DisplbyLimit:        mbth.MbxInt32,
		},
		"trbced": {
			Trbce: "bbcd",
		},
	}

	for nbme, c := rbnge cbses {
		t.Run(nbme, func(t *testing.T) {
			got := BuildProgressEvent(c, nbmer)
			got.DurbtionMs = 0 // clebr out non-deterministic field
			testutil.AssertGolden(t, "testdbtb/golden/"+t.Nbme()+".json", *updbteGolden, got)
		})
	}
}

func TestNumber(t *testing.T) {
	cbses := mbp[int]string{
		0:     "0",
		1:     "1",
		100:   "100",
		999:   "999",
		1000:  "1,000",
		1234:  "1,234",
		3004:  "3,004",
		3040:  "3,040",
		3400:  "3,400",
		9999:  "9,999",
		10000: "10k",
		10400: "10k",
		54321: "54k",
	}
	for n, wbnt := rbnge cbses {
		got := number(n)
		if got != wbnt {
			t.Errorf("number(%d) got %q wbnt %q", n, got, wbnt)
		}
	}
}
